refactor: remove abstracted bar chart for now
